# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.UserRole:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Sysware Test Services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sysware Test Services API
  version: "1.0"
paths:
...
  /role_permissions/role/{roleId}:
    delete:
      consumes:
      - application/json
      description: Delete role permissions by Role ID
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete role permissions by Role ID
      tags:
      - RolePermission
    get:
      consumes:
      - application/json
      description: Get role permissions by Role ID
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get role permissions by Role ID
      tags:
      - RolePermission
  /roles:
    get:
      consumes:
      - application/json
      description: Get all roles with pagination and search
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Order by field
        in: query
        name: order_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_by
        type: string
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Search column
        in: query
        name: column
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Role Data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - Role
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - Role
    get:
      consumes:
      - application/json
      description: Get a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get a role by ID
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role Data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update a role
      tags:
      - Role
...
  /user_roles:
    get:
      consumes:
      - application/json
      description: Get all user roles with pagination and search
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Order by field
        in: query
        name: order_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_by
        type: string
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Search column
        in: query
        name: column
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all user roles
      tags:
      - UserRole
    post:
      consumes:
      - application/json
      description: Create a new user role
      parameters:
      - description: UserRole Data
        in: body
        name: userRole
        required: true
        schema:
          $ref: '#/definitions/models.UserRole'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new user role
      tags:
      - UserRole
  /user_roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user role by ID
      parameters:
      - description: UserRole ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a user role
      tags:
      - UserRole
    get:
      consumes:
      - application/json
      description: Get a user role by ID
      parameters:
      - description: UserRole ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get a user role by ID
      tags:
      - UserRole
    put:
      consumes:
      - application/json
      description: Update a user role by ID
      parameters:
      - description: UserRole ID
        in: path
        name: id
        required: true
        type: integer
      - description: UserRole Data
        in: body
        name: userRole
        required: true
        schema:
          $ref: '#/definitions/models.UserRole'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update a user role
      tags:
      - UserRole
  /user_roles/user/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete user roles by User ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete user roles by User ID
      tags:
      - UserRole
    get:
      consumes:
      - application/json
      description: Get user roles by User ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get user roles by User ID
      tags:
      - UserRole
  /users:
    get:
      consumes:
      - application/json
      description: Get all users with pagination and search
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Order by field
        in: query
        name: order_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_by
        type: string
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Search column
        in: query
        name: column
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Update a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update a user by ID
      tags:
      - User
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"